+++ memberlist/src/main/scala/zio/memberlist/ConnectionHandler.scala
@@ -1,34 +1,18 @@
 package zio.memberlist
 
-import upickle.default.{ macroRW, _ }
 import zio._
-import zio.clock.Clock
 import zio.logging.{ log, Logging }
-import zio.memberlist.ConnectionHandler.WithPiggyback
+import zio.memberlist.MessageSink.WithPiggyback
 import zio.memberlist.encoding.ByteCodec
-import zio.memberlist.transport.{ Channel, ConnectionLessTransport }
-import zio.stream.{ Take, ZStream }
+import zio.memberlist.transport.{ Bind, Channel, ConnectionLessTransport }
+import zio.stream.Take
 
 /**
- * Represents main messages loop. responsible for receiving and sending messages.
- * @param local - local node
- * @param messages - internal queue used to process incomming messages
- * @param broadcast - broadcast for messages
- * @param transport - UDP transport
+ * Logic for connection handler for udp transport
  */
-final class ConnectionHandler(
-  val local: NodeAddress,
-  messages: Queue[Take[Error, Message[Chunk[Byte]]]],
-  broadcast: Broadcast,
-  transport: ConnectionLessTransport.Service
-) {
+object ConnectionHandler {
 
-  /**
-   * Reads message and put into queue.
-   *
-   * @param connection transport connection
-   */
-  def read(connection: Channel): IO[Error, Unit] =
+  private def read(connection: Channel, messages: Queue[Take[Error, Message[Chunk[Byte]]]]): IO[Error, Unit] =
     Take
       .fromEffect(
         connection.read >>= ByteCodec[WithPiggyback].fromChunk
@@ -51,111 +35,20 @@ final class ConnectionHandler(
       }
       .unit
 
-  private def recoverErrors[R, E, A](stream: ZStream[R, E, A]): ZStream[R with Logging, Nothing, Take[E, A]] =
-    stream.either.mapM(
-      _.fold(
-        e => log.error("error during sending", Cause.fail(e)).as(Take.halt(Cause.fail(e))),
-        v => ZIO.succeedNow(Take.single(v))
-      )
-    )
-
-  /**
-   * Sends message to target.
-   */
-  def send(msg: Message[Chunk[Byte]]): ZIO[Clock with Logging, Error, Unit] =
-    msg match {
-      case Message.NoResponse => ZIO.unit
-      case Message.BestEffort(nodeAddress, message) =>
-        for {
-          broadcast     <- broadcast.broadcast(message.size)
-          withPiggyback = WithPiggyback(local, message, broadcast)
-          chunk         <- ByteCodec[WithPiggyback].toChunk(withPiggyback)
-          nodeAddress   <- nodeAddress.socketAddress
-          _             <- transport.connect(nodeAddress).use(_.send(chunk))
-        } yield ()
-      case msg: Message.Batch[Chunk[Byte]] => {
-        val (broadcast, rest) =
-          (msg.first :: msg.second :: msg.rest.toList).partition(_.isInstanceOf[Message.Broadcast[_]])
-        ZIO.foreach_(broadcast)(send) *>
-          ZIO.foreach_(rest)(send)
-      }
-      case msg @ Message.Broadcast(_) =>
-        broadcast.add(msg)
-      case Message.WithTimeout(message, action, timeout) =>
-        send(message) *> action.delay(timeout).flatMap(send).unit
-    }
-
-  private def bind: ZManaged[Logging, TransportError, Unit] =
+  def bind(
+    local: NodeAddress,
+    transport: ConnectionLessTransport.Service,
+    messages: Queue[Take[Error, Message[Chunk[Byte]]]]
+  ): ZManaged[Logging, TransportError, Bind] =
     for {
       localAddress <- local.socketAddress.toManaged_
       _            <- log.info("bind to " + localAddress).toManaged_
       logger       <- ZManaged.environment[Logging]
-      _ <- transport
-            .bind(localAddress) { conn =>
-              read(conn)
-                .catchAll(ex => log.error("fail to read", Cause.fail(ex)).unit.provide(logger))
-            }
-    } yield ()
-
-  def process(protocol: Protocol[Chunk[Byte]]): ZIO[Clock with Logging, Nothing, Fiber.Runtime[Nothing, Unit]] = {
-    def processTake(take: Take[Error, Message[Chunk[Byte]]]) =
-      take.foldM(
-        ZIO.unit,
-        log.error("error: ", _),
-        ZIO.foreach_(_)(send(_).catchAll(e => log.error("error during send: " + e)))
-      )
-    ZStream
-      .fromQueue(messages)
-      .collectM {
-        case Take(Exit.Success(msgs)) =>
-          ZIO.foreach(msgs) {
-            case msg: Message.BestEffort[Chunk[Byte]] =>
-              Take.fromEffect(protocol.onMessage(msg))
-            case _ =>
-              ZIO.dieMessage("Something went horribly wrong.")
-          }
-      }
-      .mapMPar(10) { msgs =>
-        ZIO.foreach_(msgs)(processTake)
-      }
-      .runDrain
-      .fork *>
-      recoverErrors(protocol.produceMessages)
-        .mapMPar(10)(processTake)
-        .runDrain
-        .fork
-  }
-}
-
-object ConnectionHandler {
-
-  final case class WithPiggyback(
-    node: NodeAddress,
-    message: Chunk[Byte],
-    gossip: List[Chunk[Byte]]
-  )
-
-  implicit val codec: ByteCodec[WithPiggyback] =
-    ByteCodec.fromReadWriter(macroRW[WithPiggyback])
-
-  implicit val chunkRW: ReadWriter[Chunk[Byte]] =
-    implicitly[ReadWriter[Array[Byte]]]
-      .bimap[Chunk[Byte]](
-        ch => ch.toArray,
-        arr => Chunk.fromArray(arr)
-      )
-
-  def make(
-    local: NodeAddress,
-    broadcast: Broadcast,
-    udpTransport: ConnectionLessTransport.Service
-  ): ZManaged[Logging, TransportError, ConnectionHandler] =
-    for {
-      messageQueue <- Queue
-                       .bounded[Take[Error, Message[Chunk[Byte]]]](1000)
-                       .toManaged(_.shutdown)
-      messages = new ConnectionHandler(local, messageQueue, broadcast, udpTransport)
-      _        <- messages.bind
-    } yield messages
+      bind <- transport
+               .bind(localAddress) { conn =>
+                 read(conn, messages)
+                   .catchAll(ex => log.error("fail to read", Cause.fail(ex)).unit.provide(logger))
+               }
+    } yield bind
 
 }
